fix: keep investigate CLI under size gate
